requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.7
djangorestframework==3.14.0
idna==3.6
psycopg2-binary==2.9.9
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
stripe==8.4.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
